(No Logic) Move existing payment.component into new payment component folder
