+++ apps/web/src/app/billing/shared/adjust-payment-dialog.component.ts
@@ -10,7 +10,7 @@ import { ConfigService } from "@bitwarden/common/platform/abstractions/config/co
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { DialogService, ToastService } from "@bitwarden/components";
 
-import { PaymentComponent } from "./payment.component";
+import { PaymentComponent } from "./payment/payment.component";
 import { TaxInfoComponent } from "./tax-info.component";
 
 export interface AdjustPaymentDialogData {
