+++ apps/web/src/app/billing/shared/index.ts
@@ -1,5 +1,5 @@
 export * from "./billing-shared.module";
 export * from "./payment-method.component";
-export * from "./payment.component";
+export * from "./payment/payment.component";
 export * from "./sm-subscribe.component";
 export * from "./tax-info.component";
